requirements.txt
alembic==1.7.7
blinker==1.4
click==8.0.3
Flask==2.0.2
Flask-Migrate==3.1.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
fpdf==1.7.2
greenlet==1.1.2
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.2.0
MarkupSafe==2.0.1
Pillow==9.0.1
psycopg2-binary==2.9.3
python-dotenv==0.19.2
python-http-client==3.3.7
qrcode==7.3.1
sendgrid==6.9.7
SQLAlchemy==1.4.29
starkbank-ecdsa==2.0.3
Werkzeug==2.0.2
